More refactorings (#2449)
